I add proper dataFormat for sunrise and sunset properties and I change timeZone description for GMT

+++ MySpot/Model/MySpotData.swift
@@ -22,8 +22,8 @@ struct Temperature: Codable {
 
 struct CountryAndSys: Codable {
     let country: String
-    let sunrise: Int
-    let sunset: Int
+    let sunrise: Double
+    let sunset: Double
 }
 
 struct Weather: Codable {
